supervisor: Tolerate statically linked binaries
Support and detect statically linked leaf binaries, or ones that failed
to link, and disable shortcutting if one is encountered.

Statically linked binaries that exec() a dynamically linked one aren't
detected and aren't supported yet.

When exec()ing a statically linked binary, the exec_pending flag remains
set when the statically linked exec child quits, which is detected when
its parent wait()s for it.

At posix_spawn() we defer the new dynamically linked child if necessary
so that the posix_spawn_parent message is processed first. At the
posix_spawn_parent message we set up the same process tree as at a
fork() + exec() (with the exec_pending flag), without waiting for the
exec child. Afterwards a statically linked exec child is detected the
exact same way as with exec(): it quits with the exec_pending flag still
being set, which we catch when the parent wait()s for it.

Move sending up the posix_spawn_file_actions from the posix_spawn
message to the posix_spawn_parent message because that's when the
supervisor needs it now.

system() and popen() execute the dynamically linked /bin/sh so they
don't need special care.

Fixes #324, fixes #266, fixes #321