interceptor: Ignore stream operations without a backing fd
fmemopen and friends open streams which are backed by memory and not
by regular file descriptors. fileno() returns -1 for those and the
interceptor handled those -1-s as inherited unknown fds and reported
the issue to the supervisor.

Operations on those streams are now not reported, just performed calling
the original functions.

Fixes #787.