supervisor: Don't cache files open for writing to the blob cache
Tracking the files opened for writing is done by keeping a refcount for
each path (opened for writing) in FileName::write_fds_db_.

The refcount is incremented every time a FileFD backed by a file to be
written is constructed and the refcount is decreased when the FileFD is
destructed.

Files may be opened int a way (e.g. with O_TRUNC) that modifies the file
content at the moment the file is opened, thus the supervisor must be
notified before open() is called. In those cases a pre_open message is
sent from the interceptor and the supevisor increments the refcount in
write_fds_db_ before the intercepted process actually opens the file.

When pre_open is sent the refcount is adjusted accordingly while handling
the subsequent open in the supervisor.

Improves #778.