supervisor: Don't cache processes after the files they used changed
Introduce the concept of file (path) generations. A generation of a file
lasts until all processes that keep the file open for writing close the file.
After the number of writers decrease to zero and new process open the file for
writing a new generation is started.

The processes and the files they created or read can be cached only when the
files they used are still at the generation the process last used them
(and there are no active writers when the files are hashed or stored, but this
is ensured by earlier commits).

Fixes #739.