supervisor: Remove rarely hit wrong assert()
First it is a mistake to check fork_parent_ptr->state() instead of
potential_waiter->state() and also it is technically possible that an
exec is pending as checked a few lines above.

This assert() crashed firebuild in gcc-9's build with debugging enabled when
the supervisor detected the fork parent's exit before the orphan child's.