WIP: interceptor: Don't intercept fork() inside popen()
Glibc implementations before 2.29 implemented popen with fork()
that made firebuild intercept those forks as regular forks thanks to the
pthread_atfork handlers being called. This confused the supervisor because
it does not expect the popen()'s fork child.

See: https://sourceware.org/bugzilla/show_bug.cgi?id=17490

TODO: This is actually not a good solution. Other atfork handlers could have side effects,
and those would not be tracked properly by the supervisor. Also system() has the same problem.
TODO: Add a boolean to fork FBB messages marking if the fork was a result of popen()/system() libc implementation. A thread variable can be used to detect that in the interceptor.